+++ redis/index.js
@@ -2,11 +2,13 @@
 import client from './client';
 import getValue from './getValue';
 import setValue from './setValue';
+import setExpire from './setExpire';
 import removeValue from './removeValue';
 
 module.exports = {
     client,
     getValue,
     setValue,
+    setExpire,
     removeValue
 };